refactor highlighter to strategy pattern
